requirements.txt
python-jobspy
pyyaml
python-dotenv
schedule
requests
